tape_reset_test: Fix three position_reset checks
For the test sequence for the reset after the main "write 2 files"
tests, there are three commands that are currently failing the
test_reset_blocked_* check, as position_reset should be 1 prior to a
tape transport operation.  Therefore, reconfigure the three
test_reset_blocked_* tests prior to the rewind operation to true.

Also, refine the comment to explain the reasoning behind the status of
the test_reset_blocked_* checks.

Signed-off-by: Bryan Gurney <bgurney@redhat.com>